+++ src/microservices/Microservices.CohortPackager/Execution/ExtractJobStorage/MongoDB/MongoExtractJobStore.cs
@@ -70,16 +70,18 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB
                 .InsertOne(expectedFilesForKey);
         }
 
-        protected override void PersistMessageToStoreImpl(ExtractFileStatusMessage message, IMessageHeader header)
+        protected override void PersistMessageToStoreImpl(ExtractedFileStatusMessage message, IMessageHeader header)
         {
             if (InCompletedJobCollection(message.ExtractionJobIdentifier))
-                throw new ApplicationException("Received an ExtractFileStatusMessage for a job that is already completed");
+                throw new ApplicationException("Received an ExtractedFileStatusMessage for a job that is already completed");
 
             var newStatus = new MongoFileStatusDoc(
                 MongoExtractionMessageHeaderDoc.FromMessageHeader(message.ExtractionJobIdentifier, header, _dateTimeProvider),
-                message.AnonymisedFileName,
+                message.DicomFilePath,
+                message.OutputFilePath,
                 wasAnonymised: false,
                 isIdentifiable: true,
+                message.Status,
                 statusMessage: message.StatusMessage);
 
             _database
@@ -87,16 +89,18 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB
                 .InsertOne(newStatus);
         }
 
-        protected override void PersistMessageToStoreImpl(IsIdentifiableMessage message, IMessageHeader header)
+        protected override void PersistMessageToStoreImpl(ExtractedFileVerificationMessage message, IMessageHeader header)
         {
             if (InCompletedJobCollection(message.ExtractionJobIdentifier))
-                throw new ApplicationException("Received an IsIdentifiableMessage for a job that is already completed");
-
+                throw new ApplicationException("Received an ExtractedFileVerificationMessage for a job that is already completed");
+            
             var newStatus = new MongoFileStatusDoc(
                 MongoExtractionMessageHeaderDoc.FromMessageHeader(message.ExtractionJobIdentifier, header, _dateTimeProvider),
-                message.AnonymisedFileName,
+                message.DicomFilePath,
+                message.OutputFilePath,
                 wasAnonymised: true,
                 isIdentifiable: message.IsIdentifiable,
+                ExtractedFileStatus.Anonymised,
                 statusMessage: message.Report);
 
             _database
@@ -322,26 +326,31 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB
         protected override IEnumerable<Tuple<string, string>> GetCompletedJobAnonymisationFailuresImpl(Guid jobId)
         {
             // NOTE(rkm 2020-03-16) Files which failed anonymisation should have statuses where WasAnonymised=false and IsIdentifiable=true
-            FilterDefinition<MongoFileStatusDoc> filter = Builders<MongoFileStatusDoc>.Filter.Eq(x => x.Header.ExtractionJobIdentifier, jobId);
+            FilterDefinition<MongoFileStatusDoc> filter = FilterDefinition<MongoFileStatusDoc>.Empty;
+            filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.Header.ExtractionJobIdentifier, jobId);
             filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.WasAnonymised, false);
             filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.IsIdentifiable, true);
-
-            IAsyncCursor<MongoFileStatusDoc> cursor = _completedStatusCollection.FindSync(filter);
-            while (cursor.MoveNext())
-                foreach (MongoFileStatusDoc doc in cursor.Current)
-                    yield return new Tuple<string, string>(doc.AnonymisedFileName, doc.StatusMessage);
+            return CompletedStatusDocsForFilter(filter);
         }
 
         protected override IEnumerable<Tuple<string, string>> GetCompletedJobVerificationFailuresImpl(Guid jobId)
         {
-            FilterDefinition<MongoFileStatusDoc> filter = Builders<MongoFileStatusDoc>.Filter.Eq(x => x.Header.ExtractionJobIdentifier, jobId);
+            FilterDefinition<MongoFileStatusDoc> filter = FilterDefinition<MongoFileStatusDoc>.Empty;
+            filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.Header.ExtractionJobIdentifier, jobId);
             filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.WasAnonymised, true);
             filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.IsIdentifiable, true);
+            return CompletedStatusDocsForFilter(filter);
+        }
 
+        protected override IEnumerable<string> GetCompletedJobMissingFileListImpl(Guid jobId)
+        {
+            FilterDefinition<MongoFileStatusDoc> filter = FilterDefinition<MongoFileStatusDoc>.Empty;
+            filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.Header.ExtractionJobIdentifier, jobId);
+            filter &= Builders<MongoFileStatusDoc>.Filter.Eq(x => x.ExtractedFileStatus, ExtractedFileStatus.FileMissing);
             IAsyncCursor<MongoFileStatusDoc> cursor = _completedStatusCollection.FindSync(filter);
             while (cursor.MoveNext())
                 foreach (MongoFileStatusDoc doc in cursor.Current)
-                    yield return new Tuple<string, string>(doc.AnonymisedFileName, doc.StatusMessage);
+                    yield return doc.DicomFilePath;
         }
 
         #region Helper Methods
@@ -367,6 +376,14 @@ namespace Microservices.CohortPackager.Execution.ExtractJobStorage.MongoDB
             return _completedJobCollection
                 .Find(GetFilterForSpecificJob<MongoCompletedExtractJobDoc>(extractionJobIdentifier))
                 .SingleOrDefault() != null;
+        }
+
+        private IEnumerable<Tuple<string, string>> CompletedStatusDocsForFilter(FilterDefinition<MongoFileStatusDoc> filter)
+        {
+            IAsyncCursor<MongoFileStatusDoc> cursor = _completedStatusCollection.FindSync(filter);
+            while (cursor.MoveNext())
+                foreach (MongoFileStatusDoc doc in cursor.Current)
+                    yield return new Tuple<string, string>(doc.OutputFileName, doc.StatusMessage);
         }
 
         #endregion