+++ tests/microservices/Microservices.CohortPackager.Tests/Execution/ExtractJobStorage/ExtractJobInfoTest.cs
@@ -1,7 +1,7 @@
-﻿using System;
-using Microservices.CohortPackager.Execution.ExtractJobStorage;
+﻿using Microservices.CohortPackager.Execution.ExtractJobStorage;
 using NUnit.Framework;
 using Smi.Common.Tests;
+using System;
 
 namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
 {
@@ -48,7 +48,10 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
                 "KeyTag",
                 123,
                 "MR",
-                ExtractJobStatus.WaitingForCollectionInfo);
+                ExtractJobStatus.WaitingForCollectionInfo,
+                isIdentifiableExtraction: true,
+                isNoFilterExtraction: true
+                );
             var info2 = new ExtractJobInfo(
                 guid,
                 _dateTimeProvider.UtcNow(),
@@ -57,7 +60,10 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
                 "KeyTag",
                 123,
                 "MR",
-                ExtractJobStatus.WaitingForCollectionInfo);
+                ExtractJobStatus.WaitingForCollectionInfo,
+                isIdentifiableExtraction: true,
+                isNoFilterExtraction: true
+                );
 
             Assert.AreEqual(info1, info2);
         }
@@ -74,7 +80,10 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
                 "KeyTag",
                 123,
                 "MR",
-                ExtractJobStatus.WaitingForCollectionInfo);
+                ExtractJobStatus.WaitingForCollectionInfo,
+                isIdentifiableExtraction: true,
+                isNoFilterExtraction: true
+                );
             var info2 = new ExtractJobInfo(
                 guid,
                 _dateTimeProvider.UtcNow(),
@@ -83,7 +92,10 @@ namespace Microservices.CohortPackager.Tests.Execution.ExtractJobStorage
                 "KeyTag",
                 123,
                 "MR",
-                ExtractJobStatus.WaitingForCollectionInfo);
+                ExtractJobStatus.WaitingForCollectionInfo,
+                isIdentifiableExtraction: true,
+                isNoFilterExtraction: true
+                );
 
             Assert.AreEqual(info1.GetHashCode(), info2.GetHashCode());
         }