+++ tests/microservices/Microservices.DicomTagReader.Tests/DicomTagReaderTestHelper.cs
@@ -44,8 +44,8 @@ namespace Microservices.DicomTagReader.Tests
             // Create the test Series/Image exchanges
             Options.RabbitOptions.RabbitMqControlExchangeName = "TEST.ControlExchange";
             var tester = new MicroserviceTester(Options.RabbitOptions);
-            tester.CreateExchange(Options.DicomTagReaderOptions.ImageProducerOptions.ExchangeName, new ConsumerOptions { QueueName = TestSeriesQueueName });
-            tester.CreateExchange(Options.DicomTagReaderOptions.SeriesProducerOptions.ExchangeName, new ConsumerOptions { QueueName = TestImageQueueName });
+            tester.CreateExchange(Options.DicomTagReaderOptions.ImageProducerOptions.ExchangeName, TestSeriesQueueName );
+            tester.CreateExchange(Options.DicomTagReaderOptions.SeriesProducerOptions.ExchangeName, TestImageQueueName );
             tester.CreateExchange(Options.RabbitOptions.FatalLoggingExchange, null);
             tester.CreateExchange(Options.RabbitOptions.RabbitMqControlExchangeName, null);
             tester.Shutdown();